+++ src/instance/defaults.js
@@ -11,6 +11,7 @@ export default {
 		z: 0
 	},
 	scale: 1,
+	cleanup: true,
 	container: document.documentElement,
 	desktop: true,
 	mobile: true,